+++ src/components/assignActivity/AssignHeader.tsx
@@ -1,38 +1,19 @@
 import React from 'react';
-import { useQueryClient, useQuery, useMutation } from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
+import { useQuery } from '@tanstack/react-query';
 import { AssignData } from '@/types/assignActivityPage';
-import postAssignMyActivity from '@/api/postMyActivity';
 import queryKeys from '@/api/reactQuery/queryKeys';
-import Toast from '@/utils/Toast';
 import useCheckAssignData from '@/hooks/useCheckAssignData';
-import useMergeAssignData from '@/hooks/useMergeAssignData';
+
+import useMutationAssignData from '@/hooks/useMutateAssignData';
 
 const AssignHeader = () => {
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
+  const { assignMutation } = useMutationAssignData();
   const { checkRequireData } = useCheckAssignData();
-  const { initialAssignData } = useMergeAssignData();
   const data = useQuery({ queryKey: queryKeys.assignData() }).data as AssignData;
-
-  const mutation = useMutation({
-    mutationFn: async (assignData: AssignData) => {
-      return postAssignMyActivity(assignData);
-    },
-    onSuccess: () => {
-      Toast.success('등록 성공!!'); // 성공 시 모달 열기
-      initialAssignData(); // 쿼리 무효화
-      queryClient.invalidateQueries({ queryKey: queryKeys.activities() }); // 쿼리 무효화
-      navigate('/my/activity');
-    },
-    onError: (error) => {
-      console.error('Error:', error);
-    },
-  });
 
   const handleAssignData = async () => {
     if (checkRequireData(data)) {
-      mutation.mutate(data);
+      assignMutation.mutate(data);
     }
   };
